Add files via upload
"GeneCode Explorer" is an intricate Python project that delves into bioinformatics, specifically focusing on gene-protein translation and mutation analysis. In the first part, a script is developed to translate gene sequences into proteins, considering exon regions and mRNA-to-amino acid mappings. The second part involves a detailed analysis of amino acid compositions in the translated protein, examining their frequencies and percentages across different categories. The final part is a mutation analysis, studying the impact of a specific nucleotide change on the protein's structure and length, highlighting the correlation between genetic mutations and diseases. This project is a testament to the application of scripting in understanding complex biological processes.